Change weaken and growth methods in batch script to wait for each other
Currently its given quite a lot of incorrect fired events due to it
still looking at the server for the status and make its calculations
from that point. However due to the previous scripts not being finished
yet the information it gets is wrong resulting (mostly) in growth
scripts that get triggered way to many times.