Add small timeouts between start script methods
This is to ensure the maximum amount of ram is avaible.
Some startup script trigger script that only once to trigger
everything that is needed. This however temporarly uses more
ram then needed. To ensure all our scripts have as much ram as
possible we'll use these little timeouts to give these scripts
the time to finish.